feat [PM4-14] update parking lot navigation
